passing in enum in place of strings, changed status criteria to use enum value, newtest file for rasQueryParameters
