+++ galasa-parent/dev.galasa.framework/src/main/java/dev/galasa/framework/TestRunLifecycleStatus.java
@@ -5,6 +5,8 @@ package dev.galasa.framework;
 
 import java.util.*;
 
+import org.apache.commons.lang3.EnumUtils;
+
 public enum TestRunLifecycleStatus {
     FINISHED("finished"),
     BUILDING("building"),
@@ -17,14 +19,36 @@ public enum TestRunLifecycleStatus {
     ENDING("ending")
     ;
 
-    private String value ;
+    private String value ;  // blame Mike (He coded this on a liveshare, what is value Mike???)
 
     private TestRunLifecycleStatus(String value) {
         this.value = value ;
     }
    
-    public String toString() {
-        return value;
+
+    public TestRunLifecycleStatus getStatusFromString(String statusInput) {
+        return valueOf(statusInput);
+    }
+
+    public static boolean isStatusValid(String statusInput){
+        return EnumUtils.isValidEnum(TestRunLifecycleStatus.class, statusInput);
+    }
+
+    // @Override
+    // public String toString(){
+    //     return value.toLowerCase();
+    // }
+
+/**
+ * @return A string list from a TestRunLifecycleStatus list
+ */
+    public static List<String> convertTestRunLifecycleStausListToStringList (List<TestRunLifecycleStatus> testRunLifecycleStatuses){
+        List<String> returnList = new ArrayList<String>();
+        for (TestRunLifecycleStatus status : testRunLifecycleStatuses){
+            returnList.add(status.value);
+        }
+
+        return returnList;
     }
