+++ galasa-parent/dev.galasa.framework/src/main/java/dev/galasa/framework/spi/ras/RasSearchCriteriaStatus.java
@@ -5,13 +5,14 @@ package dev.galasa.framework.spi.ras;
 
 import javax.validation.constraints.NotNull;
 
+import dev.galasa.framework.TestRunLifecycleStatus;
 import dev.galasa.framework.spi.teststructure.TestStructure;
 
 public class RasSearchCriteriaStatus implements IRasSearchCriteria {
    
-   private final String[] statuses;
+   private final TestRunLifecycleStatus[] statuses;
    
-   public RasSearchCriteriaStatus(@NotNull String... statuses) {
+   public RasSearchCriteriaStatus(@NotNull TestRunLifecycleStatus... statuses) {
       this.statuses = statuses;
    }
    
@@ -19,24 +20,24 @@ public class RasSearchCriteriaStatus implements IRasSearchCriteria {
    public boolean criteriaMatched(@NotNull TestStructure structure) {
       
       if(structure == null) {
-         return Boolean.FALSE;
+         return false;
       }
       
       if(statuses != null) {
-         for(String status : statuses) {
-             if(status.equals(structure.getStatus())){
-                 return Boolean.TRUE;
+         for(TestRunLifecycleStatus status : statuses) {
+             if(status.toString().equals(structure.getStatus())){
+                 return true;
              }
          }
      }
       
-      return Boolean.FALSE;
+      return false;
       
       
    }
    
 
-   public String[] getStatuses() {
+   public TestRunLifecycleStatus[] getStatuses() {
        return this.statuses;
    }
